refactor to app/+fix fonts
