+++ apps/storybook/.storybook/YourTheme.ts
@@ -1,0 +1,8 @@
+import { create } from "@storybook/theming/create";
+
+export default create({
+  base: "light",
+  fontBase: '"Libre Franklin", sans-serif',
+  fontCode: "monospace",
+  brandTarget: "Polar Edge Components",
+});
