+++ apps/storybook/public/favicon.ico
