+++ package.json
@@ -18,8 +18,7 @@
   },
   "packageManager": "npm@10.8.1",
   "workspaces": [
-    "web",
-    "storybook",
+    "apps/*",
     "packages/*"
   ]
 }
