+++ packages/tailwind-config/tailwind.config.ts
@@ -4,6 +4,9 @@ import type { Config } from "tailwindcss";
 const config: Omit<Config, "content"> = {
   theme: {
     extend: {
+      fontFamily: {
+        sans: ["var(--font-libre-franklin)", '"Libre Franklin"', "sans-serif"],
+      },
       backgroundImage: {
         "glow-conic":
           "conic-gradient(from 180deg at 50% 50%, #2a8af6 0deg, #a853ba 180deg, #e92a67 360deg)",
