+++ packages/ui/package.json
@@ -36,6 +36,7 @@
     "react": "^18.2.0"
   },
   "dependencies": {
+    "@next/font": "^14.2.4",
     "@repo/tailwind-config": "^0.0.0"
   }
 }
